Move runtime code out of Next namespace
