+++ lib/graphql/execution/interpreter/handles_raw_value.rb
@@ -5,7 +5,7 @@ module GraphQL
     class Interpreter
       # Wrapper for raw values
       class RawValue
-        include GraphQL::Execution::Next::Finalizer
+        include GraphQL::Execution::Finalizer
 
         def finalize_graphql_result(query, result_data, result_key)
           result_data[result_key] = @object
