+++ lib/graphql/runtime_error.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 module GraphQL
   class RuntimeError < Error
-    include GraphQL::Execution::Next::Finalizer
+    include GraphQL::Execution::Finalizer
   end
 end
